Software culling didn't work properly in the legacy render path.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@825 01821946-851b-0410-b7f4-f1477c70e57a

+++ 2d/font.c
@@ -912,7 +912,7 @@ else {
 GrGetStringSize (s, &w, &h, &aw);
 if (!(ps->bmP = CreateStringBitmap (s, 0, 0, 0, 0, w, 1)))
 	return NULL;
-l = strlen (s) + 1;
+l = (int) strlen (s) + 1;
 if (ps->pszText && (ps->nLength < l)) {
 	D2_FREE (ps->pszText);
 	ps->nLength = ((l + 9) / 10) * 10;
